svg/pyproject.toml
[project]
name = "svg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow==11.3.0",
]
